move cumulative tests to new test_groupby_primitives file
